iio: frequency: ad916x: Fix NCO Frequency setting behavior.
Incorrect behaviours in NCO frequency setting can be due to:
1. Divisors getting truncated in 32-bit-divisor routines.
2. Truncation in computation of GCD.
3. Setting NCO frequency=0.

This patch addresses the above causes to fix the problem.

(cherry picked from commit f3d62f0f26f3143ae437553152609f76be9c1aa9)

+++ drivers/iio/frequency/ad916x/utils.h
@@ -9,7 +9,7 @@
 
 #define U64MSB 0x8000000000000000ull
 
-int adi_api_utils_gcd(int u, int v);
+int64_t adi_api_utils_gcd(int64_t u, int64_t v);
 void adi_api_utils_mult_64(uint32_t a, uint32_t b, uint32_t *hi, uint32_t *lo);
 void adi_api_utils_lshift_128(uint64_t *hi, uint64_t *lo);
 void adi_api_utils_rshift_128(uint64_t *hi, uint64_t *lo);
